Create README.md
As I was watching one of Kevin Powel's tutorials about responsive web development, I reached a point where one of the used css properties, subgrid, wasn't working in the browser that I was using , and that's because the property in question was only supported by Firefox at the time that I was following along with the tutorial, and still is as I'm currently typing this. And although using new, yet not widely supported css properties could represent a   learning experience on its own, especially when it comes to grid, I felt that it didn't perfectly align with my goal at the moment, which was to improve my capabilities in responsive web development in a sense that responds not only to different screen sizes, but also to different browsers as well.  So, instead of using Firefox in order to be able to follow along, I decided to try and develop the same layout without using the "subgrid" property, and the code before you is the result of this trial.